% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_visual.R
\name{INTERPOLACAOFUNCAO}
\alias{INTERPOLACAOFUNCAO}
\title{Interpolacao por Lagrange - Funcao:
Interpola polinomialmente pontos de uma dada funcao, e a partir da interpolacao obtida aproxima o valor da funcao em um dado ponto.}
\usage{
INTERPOLACAOFUNCAO()
}
\arguments{
\item{Funcao}{A funcao que sera usada para dar os pontos em y e entao ser feita a interpolacao}

\item{Valores_em_x}{As coordenadas em x dos pontos utilizados na interpolação, separados por espaço e em ordem crescente, exemplo: -5.58 -1 2.2 8}

\item{Pto_aproximado}{Escolha de qual ponto quer ser aproximado a partir da aproximação pela interpolação}

\item{Intervalo_x}{Define qual o intervalo do plot no eixo x, separados por espaço e na ordem crescente}

\item{Intervalo_y}{Define qual o intervalo do plot no eixo y, separados por espaço e na ordem crescente}

\item{OG_Indice}{Determina se os índices das aproximações obtidas em cada iteração serao exibidos ou não}

\item{OG_Animação}{Determina se a função interpolada sera visualizada com animação ou sera plotada de uma so vez}

\item{OG_Linhas_Verticais}{Determina se as linhas verticais ligando o ponto da função no eixo x serao exibidas ou nao}
}
\description{
Interpolacao por Lagrange - Funcao:
Interpola polinomialmente pontos de uma dada funcao, e a partir da interpolacao obtida aproxima o valor da funcao em um dado ponto.
}
