% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_visual.R
\name{TRAPEZIOS}
\alias{TRAPEZIOS}
\title{Metodo dos Trapezios:
Ilustra as iteracoes feitas pelo metodo dos trapezios para calculo de aproximacao de integral}
\usage{
TRAPEZIOS()
}
\arguments{
\item{Funcao}{A equacao que descreve a funcao para a qual se deseja calcular a integral}

\item{Intervalo}{Intervalo onde sera feita a aproximacao da integral, separado por espaco, exemplo: -3 9.87}

\item{Numero_de_Intervalos}{Define em quantas vezes o intervalo dado sera dividido}

\item{Tempo}{Regula quanto tempo durara cada iteracao}

\item{OG_Indices}{Alterna se indices dos pontos aparecerao}

\item{OG_Pintar_Area}{Alterna se a area abaixo das linhas formadas pelo trapezio devem ser pintadas ou nao}

\item{OG_Linhas_Verticais}{Alterna se as linhas ligando os pontos das divisoes com o eixo x devem ser plotados ou nao}
}
\description{
Metodo dos Trapezios:
Ilustra as iteracoes feitas pelo metodo dos trapezios para calculo de aproximacao de integral
}
