% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_visual.R
\name{INTERPOLACAOPONTOS}
\alias{INTERPOLACAOPONTOS}
\title{Interpolacao por Lagrange - Pontos:
Interpola uma funcao que passa pelos pontos dado pelo usuario, e a partir dela aproxima um ponto.}
\usage{
INTERPOLACAOPONTOS()
}
\arguments{
\item{Valores_em_x}{As coordenadas em x dos pontos utilizados na interpolacao, separados por espaco e na ordem crescente, exemplo.: -9 -3 0.58 8 22}

\item{Valores_em_y}{As coordenadas em y dos pontos utilizados na interpolacao, separados por espaco, a quantidade de termos deve ser igual a de x}

\item{Pto_aproximado}{Escolha de qual ponto quer ser aproximado a partir da aproximacao pela interpolacao}

\item{Intervalo_x}{Define qual o intervalo do plot no eixo x, separados por espaco e na ordem crescente}

\item{Intervalo_y}{Define qual o intervalo do plot no eixo y, separados por espaco e na ordem crescente}

\item{OG_Indice}{Alterna se indices dos pontos aparecerao ou nao}

\item{OG_Animacao}{Alterna se a funcao interpolada tera animacao ou sera plotada direto}

\item{OG_Linhas_Verticais}{Alterna se as linhas ligando os pontos das divisoes com o eixo x devem ser plotados ou nao}
}
\description{
Interpolacao por Lagrange - Pontos:
Interpola uma funcao que passa pelos pontos dado pelo usuario, e a partir dela aproxima um ponto.
}
