% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_visual.R
\name{SIMPSON}
\alias{SIMPSON}
\title{Método de Simpson:
Ilustra as iteracoes feitas pelo método dos Simpson, que aproxima o valor da integral de uma funcao em um dado intervalo.}
\usage{
SIMPSON()
}
\arguments{
\item{Funcao}{A equacao que descreve a funcao para a qual se deseja calcular a integral}

\item{Intervalo}{Intervalo onde sera feita a aproximacao da integral, separado por espaço, exemplo: -3 9}

\item{Numero_de_Intervalos}{Define em quantas vezes o intervalo dado sera dividido}

\item{Tempo}{Regula quanto tempo durara cada iteracao}

\item{OG_Indices}{Determina se os índices dos pontos serao exibidos}

\item{OG_Pintar_Area}{Determina se a area abaixo das curvas formadas pelo método devem ser pintadas ou nao}

\item{OG_Linhas_Verticais}{Determina se serao plotadas as linhas verticais associados a cada ponto}
}
\description{
Método de Simpson:
Ilustra as iteracoes feitas pelo método dos Simpson, que aproxima o valor da integral de uma funcao em um dado intervalo.
}
