% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_visual.R
\name{TRAPEZIOS}
\alias{TRAPEZIOS}
\title{método dos Trapezios:
Ilustra as iterações feitas pelo método dos trapezios, método que aproxima o valor da integral de uma função em um dado intervalo.}
\usage{
TRAPEZIOS()
}
\arguments{
\item{Funcao}{A equacao que descreve a funcao para a qual se deseja calcular a integral}

\item{Intervalo}{Intervalo onde sera feita a aproximacao da integral, separado por espaço, exemplo: -3 9.87}

\item{Numero_de_Intervalos}{Define em quantas vezes o intervalo dado sera dividido}

\item{Tempo}{Regula quanto tempo durara cada iteracao}

\item{OG_Indices}{Determina se os indices dos pontos serao exibidos}

\item{OG_Pintar_Area}{Determina se a area abaixo das linhas formadas pelo método devem ser pintadas ou nao}

\item{OG_Linhas_Verticais}{Determina se serao plotadas as linhas verticais associados a cada ponto}
}
\description{
método dos Trapezios:
Ilustra as iterações feitas pelo método dos trapezios, método que aproxima o valor da integral de uma função em um dado intervalo.
}
